refactor: 清理无用 cron 组件导出，迁移基础样式到 scss
