--- src/renderer/pages/cron/index.ts
@@ -1,9 +1,0 @@
-/**
- * @license
- * Copyright 2025 Sudowork (sudowork.ai)
- * SPDX-License-Identifier: Apache-2.0
- */
-
-export { default as CronJobIndicator, type CronJobStatus } from './components/CronJobIndicator';
-export { default as CronJobManager } from './components/CronJobManager';
-export { useCronJobs, useCronJobsMap } from './hooks/useCronJobs';
